pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["AskMeAgain <20632124+AskMeAgain@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
streamdeck = "^0.8.5"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
